requirements.txt
streamlit
google-genai
# additional utilities
requests